Add unit tests for data parsing
The specified Arduino plotter protocol is quite flexible, which means there multiple possible data formats which might
be supplied by a user, all of which must be supported.

In order to be able to continue the development of the project with confidence that no regressions will be introduced,
good test coverage of the data formats is essential.

Tests are added here. A GitHub Actions workflow will automatically run the tests after every relevant change as well as
periodically to catch breakage that might result from external changes.