Use node-gyp compatible version of Python in runner for test workflow
The node-gyp dependency of this project uses a Python script.

Previously, the test GitHub Actions workflow was not configured to install a specific version of Python, so whichever
version of Python 3.x that was pre-installed on the GitHub Actions runner machine was used.

The documentation for the node-gyp@7.1.2 version used by this project indicates the newest supported Python version is
3.8. Clearly newer versions did work because the workflow has been running with Python 3.10. However, the macos-latest
runner was updated to using Python 3.11 and the script now fails when `npm install` is ran in the project:

ValueError: invalid mode: 'rU' while trying to load binding.

The solution is to install a specific version of Python. It seems safest to use the newest version explicitly stated as
supported by the node-gyp@7.1.2 documentation, so Python 3.8 is installed.